dropped old database, created new db and seeded relevant records, alongside new model relationships for reviews, bookings, ticket, and payment

+++ server/Models/Review.cs
@@ -6,7 +6,7 @@ namespace server.Models;
 
 public class Review {
     public int Id { get; set; }  
-    public float Rating { get; set; } // e.g., 1 to 5 stars
+    public double Rating { get; set; } // e.g., 1 to 5 stars
     public string Comment { get; set; } 
     public DateTime ReviewDate { get; set; }  
     public int MovieId { get; set; }
